Post Controller Merged with Bubble Controller
