+++ routes/index.js
@@ -15,10 +15,8 @@ module.exports = router;
 const router = require("express").Router();
 const dashboardRoute = require("./dashboard");
 const bubbleRoute = require("./bubble");
-const postRoute = require("./post");
 
 router.use("/", dashboardRoute);
 router.use("/", bubbleRoute);
-router.use("/", postRoute);
 
 module.exports = route;
